requirements.txt
# Core
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0
asyncio-throttle>=1.0.0

# ML (price models)
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# LLM (Apple Silicon)
mlx>=0.5.0
mlx-lm>=0.10.0
huggingface-hub>=0.20.0
transformers>=4.36.0

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
matplotlib>=3.8.0

# Storage
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
rich>=13.0.0
